Rename: 일정관리 컴포넌트 폴더 변경
