+++ src/components/schedulePage/InfoModal.tsx
